Move NTCP2 handshake into submodule
